server/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework_simplejwt==5.5.0
pytesseract==0.3.13
gunicorn==20.1.0
pytest==8.3.5